% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATEs_utils.R
\name{nuisance_cf_grf}
\alias{nuisance_cf_grf}
\title{This function creates the nuisance parameters p(x), mu(x), and mu_d(x)
via cross-fitting using the \code{\link{grf}} package}
\usage{
nuisance_cf_grf(y, d, x, index, args_p = list(), args_y = list(),
  args_y0 = list(), args_y1 = list())
}
\arguments{
\item{y}{Vector of outcome values}

\item{d}{Vector of treament indicators}

\item{x}{Matrix of covariates (N x p matrix)}

\item{index}{List indicating indices for cross-fitting (e.g. obtained by \code{createFolds} of \code{\link{caret}} pkg)}

\item{args_p}{List of arguments passed to estimate propensity score model}

\item{args_y}{List of arguments passed to estimate outcome model}

\item{args_y0}{List of arguments passed to estimate outcome model of non-treated}

\item{args_y1}{List of arguments passed to estimate outcome model of treated}
}
\value{
Returns n x 4 matrix containing the nuisance parameters
}
\description{
This function creates the nuisance parameters p(x), mu(x), and mu_d(x)
via cross-fitting using the \code{\link{grf}} package
}
