% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATEs_utils.R
\name{cf_dml1}
\alias{cf_dml1}
\title{This function implements the 50:50 cross-fitting}
\usage{
cf_dml1(est, y, d, x, np, xnew, index, args_tau = list())
}
\arguments{
\item{est}{Vector of outcome values}

\item{d}{Vector of treament indicators}

\item{x}{Matrix of covariates (N x p matrix)}

\item{index}{List indicating indices for cross-fitting (e.g. obtained by \code{createFolds} of \code{\link{caret}} pkg)}

\item{args_p}{List of arguments passed to estimate propensity score model}

\item{args_y}{List of arguments passed to estimate outcome model}

\item{args_y1}{List of arguments passed to estimate outcome model of treated}

\item{args_y0}{List of arguments passed to estimate outcome model of non-treated}
}
\value{
Returns n x 4 matrix containing the nuisance parameters
#'
}
\description{
This function implements the 50:50 cross-fitting
}
