% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATEs.R
\name{IATEs}
\alias{IATEs}
\title{This function produces the predicted values of all 13 estimator considered in KLS}
\usage{
IATEs(y_t, d_t, x_t, tau_t, x_v)
}
\arguments{
\item{y_t}{Vector of training outcome values}

\item{d_t}{Vector of training treament indicators}

\item{x_t}{Matrix of training covariates (N x p matrix)}

\item{x_v}{Matrix of validation covariates (N x p matrix)}
}
\value{
Returns n x 13 matrix containing the IATEs of the validation sample
}
\description{
This function produces the predicted values of all 13 estimator considered in KLS
}
